Avance subida de archivos por servicio ultra zócalo

+++ app/src/main/java/com/naposystems/napoleonchat/repository/conversation/ConversationRepository.kt
@@ -627,7 +627,7 @@ class ConversationRepository @Inject constructor(
                                     offer(
                                         DownloadAttachmentResult.Progress(
                                             itemPosition,
-                                            finalPercentage
+                                            finalPercentage.toFloat()
                                         )
                                     )
                                 }
